requirements.txt
aioredis==1.3.1
async-timeout==3.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
fastapi-utils==0.2.1
h11==0.12.0
hiredis==1.1.0
idna==2.10
motor==2.3.1
pydantic==1.8
pymongo==3.11.3
python-dotenv==0.15.0
requests==2.25.1
sqlalchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
